Refactor into separate modules for work extraction, analysis, glue
