+++ covid_moonshot/__init__.py
@@ -1,0 +1,1 @@
+from .lib import *
